Apply confirm

+++ src/app/project/containers/config/config.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core'
 import { Router } from '@angular/router'
 import { Store } from '@ngrx/store'
 import { ProjectService } from 'app/project/services/project.service'
+import { ConfirmService } from 'app/shared/services/confirm.service'
 import * as projectsAction from 'app/core/actions/projects.action'
 import * as fromCore from 'app/core/reducers'
 import * as fromProject from 'app/project/reducers'
@@ -20,6 +21,7 @@ export class ConfigComponent implements OnInit {
   constructor (
     private store: Store<any>,
     private projectService: ProjectService,
+    private confirmService: ConfirmService,
     private router: Router
   ) {
     this.store.select(fromCore.getProjectsItems)
@@ -39,7 +41,7 @@ export class ConfigComponent implements OnInit {
     this.project.environments = entities
   }
 
-  onUpdate () {
+  update () {
     const payload = {
       ...this.project,
       environments: json.toJSON(this.project.environments)
@@ -59,19 +61,26 @@ export class ConfigComponent implements OnInit {
       })
   }
 
-  onDelete () {
-    this.projectService.delete(this.project.id)
-      .subscribe(res => {
-        if (!res.error) {
-          this.projects = this.projects.filter(project => {
-            return project.id !== this.project.id
-          })
-          this.store.dispatch(new projectsAction.ItemsAction(this.projects))
-          if (this.projects.length > 0) {
-            this.router.navigateByUrl(`/project/${this.projects[0].id}`)
-          } else {
-            this.router.navigateByUrl(`/my-account`)
-          }
+  delete () {
+    this.confirmService.open({
+      message: 'Are you sure you want to delete this project'
+    })
+      .afterClose(val => {
+        if (val) {
+          this.projectService.delete(this.project.id)
+            .subscribe(res => {
+              if (!res.error) {
+                this.projects = this.projects.filter(project => {
+                  return project.id !== this.project.id
+                })
+                this.store.dispatch(new projectsAction.ItemsAction(this.projects))
+                if (this.projects.length > 0) {
+                  this.router.navigateByUrl(`/project/${this.projects[0].id}`)
+                } else {
+                  this.router.navigateByUrl(`/my-account`)
+                }
+              }
+            })
         }
       })
   }
